LayoutTests/fast/js/sputnik/Conformance/08_Types/8.2_The_Null_Type/S8.2_A2-expected.txt

CONSOLE MESSAGE: line 76: SyntaxError: Parse error
S8.2_A2

PASS Expected parsing failure

TEST COMPLETE
